NDAS/ndas/requirements.txt
pyqtgraph==0.11.1
numpy==1.18.4
hickle==4.0.4
wfdb==3.2.0
scipy==1.5.3
pandas==1.0.3
PyQt5==5.15.4
PyYAML==5.4.1
